+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # _____ Change Log
 
+- refactored item flags, see README.md
+- add default params for items
+- add type checks for given params
+- removed recursive deletion in safeDown(). These migrations cannot be reverted. 
+- deprecate _exists and _force item params
+
 ## 1.0.0 under development
 
 - Initial release.